Use --locked with cargo install.
The diesel cli does not currently have a lockfile checked in, but it
might in the future and/or other tools with a lockfile might be added to
this projcet in the future.

Using --locked is beneficial because it increases the reproducibility of
CI runs and increases the liklihood that the cache will be useful.