Refactor ZeroizedPassword into CleartextPassword
The type represents a cleartext password; the fact that it needs
to be zeroized is derived from the fact that it is cleartext.
Rename it to be more accurate in what the type represents instead of
refering to one implementation detail.

Protect the member so it can't be arbitrarily changed.