Move app builder into macro.
This is the first step of refactoring so that tests and main both use
the same set of services. It is in a macro instead of a helper function
because the builder returns an App<AppService>, and AppService is hidden
inside a private module. This means that the return type of a helper
function cannot be specified, so a macro is used to re-write the builder
everywhere instead.